Avoid re-running gom all the time
The make dependency on `_vendor/src/$(IMPORT_PATH)` seems like it should work,
but in fact make de-references the symlink and looks at the mtime of the
symlink target to determine whether the `_vendor` task should be re-run. The
symlink target is the project root directory, but that changes whenever the
output binary is re-built (ie. quite often). The net result is that gom install
is run more frequently than it needs to be.

The workaround here is to add a hidden file indicating when the symlink was
created, rather than relying on the symlink itself.