Read entire post body
We were seeing intermittent errors in the client because the TCP connection was
being closed before the response body had been sent back to the client. We
suspect that this was due to the fact that sometimes the go client would not
read the entire post body for larger requests, and nginx was getting upset that
the upstream had stopped reading and therefore closed the connection before the
upstream had a chance to respond.

We have worked around this by using ioutil.ReadAll to read the entire post body.

Note that this means that the entity extractor will read an unlimited post body
size. However, in practice the amount of data that can be sent to it is limited
by nginx proxies in front.